+++ src/app/(afterLogin)/(main)/mypage/account-deletion/page.tsx
@@ -1,31 +1,19 @@
 'use client';
 
-import Modal from '@/app/_components/common/modal/Modal';
-import { MODAL_TYPE_MAP } from '@/app/_constants/modal';
-import { useRouter, useSearchParams } from 'next/navigation';
+import dynamic from 'next/dynamic';
+import React from 'react';
+
+const AccountDeletionContent = dynamic(
+  () => import('@/app/(afterLogin)/(main)/mypage/account-deletion/_components/AccountDeletionContent'),
+  {
+    ssr: false,
+  },
+);
 
 export default function AccountDeletion() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const targetType = searchParams.get('targetType') || 'member';
-
-  const content = MODAL_TYPE_MAP[targetType];
-
-  const handleCancelClick = () => {
-    router.back();
-  };
-
-  const handleDeletionConfirm = () => {};
-
   return (
-    <div className="absolute">
-      <Modal
-        title="삭제하기"
-        content={content}
-        buttonText="탈퇴"
-        onCancel={handleCancelClick}
-        onExecute={handleDeletionConfirm}
-      />
+    <div>
+      <AccountDeletionContent />
     </div>
   );
 }
